Refactor: navBar 코드 리팩토링 #60
